platform/backend/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "v-platform"
version = "0.1.0"
description = "Reusable platform framework: auth, RBAC, audit, user management, UI Kit"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "sqlalchemy>=2.0.25",
    "psycopg2-binary>=2.9.9",
    "redis>=5.0.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "cryptography>=42.0.0",
    "structlog>=24.1.0",
    "slowapi>=0.1.9",
    "prometheus-client>=0.19.0",
    "aiosmtplib>=3.0.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.6",
    "email-validator>=2.1.0",
]

[project.optional-dependencies]
dev = ["ruff", "pytest", "pytest-asyncio"]

[tool.setuptools.packages.find]
include = ["v_platform*"]
